consumer/requirements.txt
kafka-python
polars
boto3
pathlib


